requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]==4.8
pymongo
scikit-learn
dill
pyaml
mlflow
fastapi
uvicorn
dagshub
##-e .